requirements.txt
numpy
tensorpack
opencv
pyarrow
atari-py
gym
absl-py
